# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse_EntityResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EntityResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[EntityResponse]",
        "type": "object"
      },
...
      "ApiResponse_RelationshipResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RelationshipResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[RelationshipResponse]",
        "type": "object"
      },
...
      "ApiResponse_list_EntityResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[EntityResponse]]",
        "type": "object"
      },
...
      "ApiResponse_list_RelationshipResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RelationshipResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "status": {
            "$ref": "#/components/schemas/ResponseStatus",
            "default": "success"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[RelationshipResponse]]",
        "type": "object"
      },
...
      "EntityAttributesUpdateRequest": {
        "description": "Replacement descriptive attributes for an entity.",
        "properties": {
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attributes",
            "type": "object"
          }
        },
        "required": [
          "attributes"
        ],
        "title": "EntityAttributesUpdateRequest",
        "type": "object"
      },
      "EntityCreateRequest": {
        "description": "Client-supplied fields for creating (or reusing) an entity.",
        "properties": {
          "aliases": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Attributes",
            "type": "object"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "display_name",
          "confidence",
          "source"
        ],
        "title": "EntityCreateRequest",
        "type": "object"
      },
      "EntityResponse": {
        "description": "Flat projection of an entity.",
        "properties": {
          "aliases": {
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "attributes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Attributes",
            "type": "object"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "display_name",
          "confidence",
          "source",
          "attributes",
          "aliases"
        ],
        "title": "EntityResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RelationshipConfidenceUpdateRequest": {
        "description": "New confidence for an existing relationship.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          }
        },
        "required": [
          "confidence"
        ],
        "title": "RelationshipConfidenceUpdateRequest",
        "type": "object"
      },
      "RelationshipCreateRequest": {
        "description": "Client-supplied fields for creating a relationship.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "source_entity_id": {
            "title": "Source Entity Id",
            "type": "string"
          },
          "supporting_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          },
          "target_entity_id": {
            "title": "Target Entity Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "source_entity_id",
          "target_entity_id",
          "type",
          "confidence",
          "supporting_evidence"
        ],
        "title": "RelationshipCreateRequest",
        "type": "object"
      },
      "RelationshipResponse": {
        "description": "Flat projection of a relationship.",
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source_entity_id": {
            "title": "Source Entity Id",
            "type": "string"
          },
          "supporting_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Supporting Evidence",
            "type": "array"
          },
          "target_entity_id": {
            "title": "Target Entity Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_entity_id",
          "target_entity_id",
          "type",
          "confidence",
          "supporting_evidence",
          "created_at"
        ],
        "title": "RelationshipResponse",
        "type": "object"
      },
...
      "ResponseMeta": {
        "description": "Response metadata common to success and error responses.",
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "correlation_id",
          "timestamp",
          "api_version"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
      "ResponseStatus": {
        "description": "The outcome status carried by every API response.",
        "enum": [
          "success",
          "error"
        ],
        "title": "ResponseStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SentinelAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/graph/entities": {
      "post": {
        "operationId": "create_entity_api_v1_graph_entities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EntityResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Entity",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/entities/{entity_id}": {
      "delete": {
        "description": "Erase a person-linked entity \u2014 the right-to-be-forgotten path.\n\nRequires the ``knowledge:erase`` capability (ADR-019), enforced by the\nauthorization seam before this endpoint runs.\n\n*Which* entities are person-linked is never inferred from data content:\nthe caller names the entity. The node keeps its identifier so incident\nrelationships still resolve to an explicit erased node. Idempotent.",
        "operationId": "erase_entity_api_v1_graph_entities__entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EntityResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Erase Entity",
        "tags": [
          "graph"
        ]
      },
      "get": {
        "operationId": "get_entity_api_v1_graph_entities__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EntityResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entity",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/entities/{entity_id}/attributes": {
      "post": {
        "operationId": "update_entity_attributes_api_v1_graph_entities__entity_id__attributes_post",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityAttributesUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_EntityResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Entity Attributes",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/entities/{entity_id}/neighbors": {
      "get": {
        "operationId": "find_neighbors_api_v1_graph_entities__entity_id__neighbors_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "depth",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Max Nodes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_EntityResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Find Neighbors",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/entities/{entity_id}/relationships": {
      "get": {
        "operationId": "list_relationships_for_entity_api_v1_graph_entities__entity_id__relationships_get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_RelationshipResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Relationships For Entity",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/relationships": {
      "post": {
        "operationId": "create_relationship_api_v1_graph_relationships_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RelationshipResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Relationship",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/relationships/{relationship_id}": {
      "get": {
        "operationId": "get_relationship_api_v1_graph_relationships__relationship_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "relationship_id",
            "required": true,
            "schema": {
              "title": "Relationship Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RelationshipResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Relationship",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/relationships/{relationship_id}/confidence": {
      "post": {
        "operationId": "update_relationship_confidence_api_v1_graph_relationships__relationship_id__confidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "relationship_id",
            "required": true,
            "schema": {
              "title": "Relationship Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipConfidenceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RelationshipResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Relationship Confidence",
        "tags": [
          "graph"
        ]
      }
    },
...
  }
}